+++ app/src/main/java/com/example/plantmonitoringsystem/Fragments/FragmentTemperature.java
@@ -23,7 +23,6 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.ArrayList;
-import java.util.BitSet;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -34,7 +33,7 @@ public class FragmentTemperature extends Fragment {
     private DatabaseReference reference;
     private ListView listView;
     private String zone;
-    private ArrayList<Integer> count = new ArrayList<Integer>();
+    private ArrayList<Long> count = new ArrayList<Long>();
 
     public FragmentTemperature(String Zone) {
         // Required empty public constructor
@@ -62,7 +61,7 @@ public class FragmentTemperature extends Fragment {
 
                 //get the count
                 count.clear();
-                for(int i=0;i<params.size();i++){
+                for(long i=0;i<params.size();i++){
                     count.add(i);
                 }
